Add support for individual user configuration
This commit adds support for overriding the global
configuration through ~/.ssmtprc. When this file is
found the config file in /etc is not used.

It also adds a new keywork "EmailAddress" that is only
valid in the individual configuration file and sets the
address to be used in the From: field of messages sent
by the user.

Man pages are modified to reflect the new feature.