Added Message Queueing Support
This commits adds support for saving failed messages
to a queue for later sending. This is a work in progress.
At the moment it just saves the email to a directory
specified by the QueueDir config option using the
following filename:

 <RECIPIENT>-XXXXXX

If a message has multiple recipients then multiple copies
will be saved.