Retought queue implementation
Reworked the queue implementation. Now all every queued
email is saved only once and the first line contains a
comma separated list of recipients. When processing the
queue we strip that line before piping it to a child and
pass the recipients as arguments. We'll also be using the
same executable for the agent. The emails are saved as
mail-XXXXXX but that's meaningless and will probably
change. It needs a lot of cleanup but it works.

TODO:
1. It's no longer necessary to buffer the whole message
   before writing to disk since we'll only save one copy.

2. Don't use readline() when processing queue to be sure
   it works anywhere the original ssmtp works.